JPC: create new stateless component to inject extra info section to plans pages.
This removes code repetition in defining the extra component both in plans-landing
page for unconected sites, ands plans.jsx for connected ones.
